Modified Genomes to be UClass and potential fixes for it
Genomes has been modified to be a UClass inheriting from UObject again, but any instances of Genome are created once only when the thread starts in order to prevent Garbage Collection being at odds with it.

Issues have arrived coming from some items in TArray<UGenomes*> becoming NULL or the Bits associated within one comes back Invalid or Empty. Possibly to do with Garbage Collection, but unsure as these have UPROPERTY() to prevent being collected.

The GA Module's usages have been commented out of the GameMode to prevent it being used whilst other areas are being built as to not conflict. Will be restored when potential fixes are applied.