GA NULL fix and start of BP Weapons
A fix has been put in place for the GAFunctionality and GAController becoming NULL after being created in the GameMode constructor (by moving to BeginPlay).

Beginnings of BPWeapons being put in place for game functionality.

Now have separate bug: Newly created enemies & generations are not being passed up into the thread. These are being added to the lists within the thread but not being read by dequeues. Possible fix is to move the dequeue calls as delegates to be fired elsewhere, rather than in the tick component of GAController.