Cleaned some TODOs up and implemented duplicate fit enemy check
Cleaned out some TODOs that were left in the GA. Some future TODOs have been added to track progress.

Duplicate new fit enemies will now be filtered out as to not result in duplicates in NewEnemies.

Need to check to ensure that the randomness is mostly random on first-pass for creating genomes, as they mostly resort to all 0/false when starting, resulting in all enemies having the same stats.