Started cleanup and making items private for ease of use
Almost all of the Logging that isn't intrusive/display has been commented out for the time being. Functions/variables in the GAController should be placed as private unless usable by User. Queues must be public for now until they are reworked to not require public (must to prevent user interference).