Thread Shutdown fixed
Can finally interrupt and shut down the GA thread whilst it is still running! Added an explicit call to GAController->EndPlay() from within the GameMode's Endplay function, since it was having trouble calling it itself. Works like a dream and completely shuts it down without crashes or issues when unexpectedly finishing before the thread completes.

Also added a few content packs for visualization of the changes made by the algorithm, to better represent what it does to future enemies. May change out for a better or custom model, but as this one was free and (mostly) should work, it's fine for now.